Update to,
1. Remove test_utils and dependency from routing src.
2. Move churn_node, node and local_network binaries in tests folder to examples folder.
3. Catch 'error' NotEnoughSignatures before printing.
4. Add local_network integration test target.
