+++ examples/utils/mod.rs
@@ -15,27 +15,13 @@
 // Please review the Licences for the specific language governing permissions and limitations
 // relating to use of the SAFE Network Software.
 
-/// Simulate a vault.
+/// Node.
 pub mod node;
-/// Simulate a client.
-pub mod client;
-/// Create a node that churns.
+/// ChurnNode.
 pub mod churn_node;
-/// Various random values for types available to tests.
-pub mod messages_util;
+/// Client.
+pub mod client;
 
-pub use self::node::*;
-pub use self::client::*;
-pub use self::churn_node::*;
-pub use self::messages_util::*;
-
-use xor_name::XorName;
-
-/// Return XOR of two XorName's.
-pub fn xor(lhs: &XorName, rhs: &XorName) -> XorName {
-    let mut result = XorName::new([0u8; 64]);
-    for i in 0..lhs.0.len() {
-        result.0[i] = lhs.0[i] ^ rhs.0[i];
-    }
-    result
-}
+pub use self::node::Node;
+pub use self::churn_node::ChurnNode;
+pub use self::client::Client;
